+++ packages/claims-controller/src/index.ts
@@ -42,9 +42,17 @@ export type {
   ClaimsServiceGetClaimsApiUrlAction,
   ClaimsServiceGetClaimByIdAction,
   ClaimsServiceGenerateMessageForClaimSignatureAction,
+  ClaimsServiceMethodActions,
+} from './ClaimsService-method-action-types.js';
+
+export type {
   ClaimsServiceActions,
   ClaimsServiceConfig,
   ClaimsServiceMessenger,
+  ClaimsServiceInvalidateQueriesAction,
+  ClaimsServiceCacheUpdatedEvent,
+  ClaimsServiceGranularCacheUpdatedEvent,
+  ClaimsServiceEvents,
 } from './ClaimsService.js';
 
 export {