+++ packages/wallet/src/Wallet.test.ts
@@ -188,7 +188,7 @@ describe('Wallet', () => {
 
     const results = await wallet.init();
 
-    expect(results).toHaveLength(2);
+    expect(results).toHaveLength(3);
   });
 
   it('disallows modifying the messenger', async () => {