+++ packages/wallet/src/initialization/instances/claims-controller/types.ts
@@ -1,15 +1,6 @@
-import type {
-  ClaimsControllerOptions,
-  ClaimsServiceConfig,
-} from '@metamask/claims-controller';
+import type { ClaimsControllerOptions } from '@metamask/claims-controller';
 
 export type ClaimsControllerInstanceOptions = Omit<
   ClaimsControllerOptions,
   'messenger' | 'state'
 >;
-
-export type ClaimsServiceInstanceOptions = Omit<
-  ClaimsServiceConfig,
-  'messenger' | 'env'
-> &
-  Partial<Pick<ClaimsServiceConfig, 'env'>>;