+++ packages/wallet/src/types.ts
@@ -6,10 +6,8 @@ import type {
   RootMessenger,
 } from './initialization/defaults.js';
 import type { ApprovalControllerInstanceOptions } from './initialization/instances/approval-controller/types.js';
-import type {
-  ClaimsControllerInstanceOptions,
-  ClaimsServiceInstanceOptions,
-} from './initialization/instances/claims-controller/types.js';
+import type { ClaimsControllerInstanceOptions } from './initialization/instances/claims-controller/types.js';
+import type { ClaimsServiceInstanceOptions } from './initialization/instances/claims-service/types.js';
 import type { ConnectivityControllerInstanceOptions } from './initialization/instances/connectivity-controller/types.js';
 import type { GasFeeControllerInstanceOptions } from './initialization/instances/gas-fee-controller/types.js';
 import type { KeyringControllerInstanceOptions } from './initialization/instances/keyring-controller/types.js';